// v2.client/src/shared/api/generated/models/file-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Privilage } from './privilage';

export interface FileUserDto {
    'userId': string | null;
    'login'?: string | null;
    'privilage'?: Privilage;
}